✨ Working refactor3
